Update beijing  times

+++ Core/Inc/wifi_fun.h
@@ -33,7 +33,7 @@ typedef struct _WIFI_FUN{
     uint8_t soft_ap_config_flag;
     uint8_t get_rx_beijing_time_flag;
 	uint8_t rx_beijing_decode_flag;
-	uint8_t gTimer_publish_times;
+	uint16_t gTimer_publish_times;
 	uint8_t wifi_reconnect_read_flag;
